pyproject.toml
# SPDX-FileCopyrightText: 2023 Phu Hung Nguyen <phuhnguyen@outlook.com>
# SPDX-License-Identifier: CC0-1.0

[tool.poetry]
name = "mdit-py-hugo"
version = "0.3.1"
description = "Collection of markdown-it-py plugins and extensions that add supports for Hugo syntaxes and functionalities"
authors = ["Phu Hung Nguyen <phuhnguyen@outlook.com>"]
license = "LGPL-2.1-or-later"
readme = "README.md"
repository = "https://github.com/phunh/mdit-py-hugo"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3"
]
packages = [{include = "mdit_py_hugo"}]

[tool.poetry.dependencies]
python = "^3.8"
markdown-it-py = "^3.0.0"
mdit-py-plugins = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[flake8]
max-line-length = 120
